[Bronze V] Title: 이상한 기호, Time: 120 ms, Memory: 14212 KB -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -8,7 +8,7 @@ public class Main {
 
     public static void main(String[] args) throws IOException {
         StringTokenizer st = new StringTokenizer(br.readLine());
-        long A = Integer.parseInt(st.nextToken()), B = Integer.parseInt(st.nextToken());
+        int A = Integer.parseInt(st.nextToken()), B = Integer.parseInt(st.nextToken());
         bw.append(String.valueOf((A + B) * (A - B)));
         bw.flush();
         bw.close();
